Remove auth package
