--- packages/auth/.gitignore
@@ -1,3 +1,0 @@
-tsconfig.tsbuildinfo
-dist/*
-sessions.sqlite3
