--- packages/auth/lib/index.ts
@@ -1,4 +1,0 @@
-import authServer from './auth.server.js';
-import { getOauthCallbackUrl, getPort } from './utils/utils.js';
-
-export { authServer, getOauthCallbackUrl, getPort };
