--- packages/auth/tsconfig.json
@@ -1,8 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "rootDir": "lib",
-        "outDir": "dist"
-    },
-    "include": ["lib/**/*"]
-}
