+++ packages/server/lib/clients/auth.client.ts
