+++ packages/server/lib/utils/analytics.ts
