+++ packages/server/lib/utils/web-socket-error.ts
