+++ packages/server/package.json
@@ -19,6 +19,45 @@
         "npm": ">=6.14.11"
     },
     "dependencies": {
-        "@nangohq/auth": "^0.9.0"
+        "@sentry/node": "^7.37.2",
+        "@sentry/tracing": "^7.37.2",
+        "braintree": "^3.13.0",
+        "connect-sqlite3": "^0.9.13",
+        "cookie-parser": "^1.4.6",
+        "cors": "^2.8.5",
+        "dotenv": "^16.0.3",
+        "express": "^4.18.2",
+        "express-session": "^1.17.3",
+        "ip": "^1.1.8",
+        "js-yaml": "^4.1.0",
+        "knex": "^2.3.0",
+        "oauth": "^0.10.0",
+        "passport": "^0.6.0",
+        "passport-local": "^1.0.0",
+        "pg": "^8.8.0",
+        "posthog-node": "^2.3.0",
+        "simple-oauth2": "^5.0.0",
+        "uuid": "^9.0.0",
+        "winston": "^3.8.2",
+        "ws": "^8.12.1"
+    },
+    "devDependencies": {
+        "@types/braintree": "^3.3.6",
+        "@types/connect-sqlite3": "^0.9.2",
+        "@types/cookie-parser": "^1.4.3",
+        "@types/cors": "^2.8.12",
+        "@types/csurf": "^1.11.2",
+        "@types/express": "^4.17.13",
+        "@types/express-session": "^1.17.6",
+        "@types/ip": "^1.1.0",
+        "@types/js-yaml": "^4.0.5",
+        "@types/node": "^18.7.6",
+        "@types/oauth": "^0.9.1",
+        "@types/passport": "^1.0.12",
+        "@types/passport-local": "^1.0.35",
+        "@types/simple-oauth2": "^4.1.1",
+        "@types/uuid": "^8.3.4",
+        "@types/ws": "^8.5.4",
+        "typescript": "^4.7.4"
     }
 }
