+++ packages/server/tsconfig.json
@@ -4,6 +4,5 @@
         "rootDir": "lib",
         "outDir": "dist"
     },
-    "references": [{ "path": "../auth" }],
     "include": ["lib/**/*"]
 }
