+++ webapp/package.json
@@ -6,7 +6,8 @@
         "dev": "next dev",
         "build": "next build",
         "start": "next start",
-        "lint": "next lint"
+        "lint": "next lint",
+        "export": "next build && next export"
     },
     "dependencies": {
         "@next/font": "13.1.6",
